solution

+++ restaurants/views.py
@@ -3,6 +3,7 @@ from .models import Restaurant, Item
 from .forms import RestaurantForm, ItemForm, SignupForm, SigninForm
 from django.contrib.auth import login, authenticate, logout
 
+
 def signup(request):
     form = SignupForm()
     if request.method == 'POST':
@@ -59,6 +60,9 @@ def restaurant_detail(request, restaurant_id):
     return render(request, 'detail.html', context)
 
 def restaurant_create(request):
+    if not request.user.is_authenticated:
+        return redirect('signin')
+
     form = RestaurantForm()
     if request.method == "POST":
         form = RestaurantForm(request.POST, request.FILES)
@@ -73,8 +77,10 @@ def restaurant_create(request):
     return render(request, 'create.html', context)
 
 def item_create(request, restaurant_id):
-    form = ItemForm()
     restaurant = Restaurant.objects.get(id=restaurant_id)
+    if not (request.user == restaurant.owner or request.user.is_staff):
+        return redirect('no-access')
+    form = ItemForm()
     if request.method == "POST":
         form = ItemForm(request.POST)
         if form.is_valid():
@@ -89,7 +95,10 @@ def item_create(request, restaurant_id):
     return render(request, 'item_create.html', context)
 
 def restaurant_update(request, restaurant_id):
+    
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
+    if not (request.user == restaurant_obj.owner or request.user.is_staff):
+        return redirect('no-access')
     form = RestaurantForm(instance=restaurant_obj)
     if request.method == "POST":
         form = RestaurantForm(request.POST, request.FILES, instance=restaurant_obj)
@@ -104,5 +113,14 @@ def restaurant_update(request, restaurant_id):
 
 def restaurant_delete(request, restaurant_id):
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
+    if not request.user.is_staff:
+        return redirect('no-access')
     restaurant_obj.delete()
     return redirect('restaurant-list')
+
+def no_access(request):
+    context = {
+
+    }
+
+    return render(request, 'no_access.html', context)
